backend/requirements.txt
# VeriClip AI - Production Dependencies
# Core web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for validation
pydantic==2.9.2
pydantic-settings==2.13.1

# Google Cloud & AI
firebase-admin==6.4.0
google-cloud-aiplatform==1.42.0
google-genai==0.3.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.19.0
google-cloud-videointelligence==2.12.0

# Cryptography & C2PA
cryptography==46.0.7
cbor2==5.6.2

# HTTP client
httpx==0.27.2

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.2.2
